requirements.txt
streamlit==1.32.0
opencv-python-headless
numpy
Pillow
streamlit-cropper
google-genai
pyairtable
